MSI-810: Exception should be thrown when Stock Source Link related to Default Stock is saved or deleted.
